Align Surface lab: archive unreviewed stories, lock LayeredPaper colors, merge ghost-spark stories
- Move SkeuomorphicCard, WarmthCurvesAtScale, PaperModels, TopBarTreatments,
  FrameRecess out of North Star §1 Surface System into Lab/Archive/Surface,
  dropping the "(unreviewed)" name suffix now that they're archived. §1 keeps
  only the three locked stories: SurfaceRampSystem, LayeredPaper, PaperWithBrand.
- Fix LayeredPaper to render with the locked surface colors instead of ad-hoc
  exploratory tones: neutral paper sheets now derive from deriveSurfaceRamp()
  (base #252321 / elevated #2A2827 / raised #2D2C2B), and the accent sheet uses
  the categorical dark-variant brand orange -600 (#B94A00). Desk background
  (#242019) was already correct. Paper treatment (grain/rim/shadow/offset)
  unchanged.
- Merge PhaseMarks.stories.tsx into GhostSpark.stories.tsx — phase marks are
  the ghost-spark's phase-colored zero-axis, not a separate component. Lab/
  Components now has one file per real component: Ghost Spark, Verdict Hero,
  Velocity Hero.
